Create BaseActivity for SharedPreferences and OptionsMenu
Moved all SharedPreferences and OptionsMenu functionality to BaseActivity. This Activity can now be extended by all other activities to give access to the logged in userId method as well as having automatic creation of the functional options menu.
