+++ app/src/main/AndroidManifest.xml
@@ -15,6 +15,9 @@
         android:theme="@style/Theme.KitchInventory"
         tools:targetApi="31">
         <activity
+            android:name=".activities.BaseActivity"
+            android:exported="false" />
+        <activity
             android:name=".activities.AdminActivity"
             android:exported="false" />
         <activity
